Add OrderItem and RestaurantMenuItem, update Order and OrderController
This commit adds two new models, OrderItem and RestaurantMenuItem. The Order model is updated to include a list of OrderItems, enabling the application to keep track of each item in an order. The OrderController is updated with new functionality allowing users to get a list of their orders and view restaurant offers. The DataBaseAPI controller was also restructured for better readability, added Javadoc comments to methods and renamed some methods for consistency and clarity. Aim of changes is to facilitate online order placements and tracking.